use strings as immutable default arguments (#162)
Fixes #161 

Avoid using mutable objects as default arguments as functions. Even
though this particular usage was safe, it opens up the projects to
issues down the road if the default argument is mutated within the
function. For this usage, we don't need an actual `list` object, just a
sequence of strings, so we can use a string itself.